aromatic rings are not identified correctly in ringAromMap

+++ include/MCSCompound.h
@@ -169,8 +169,9 @@ namespace FMCS {
         }
         
         //saving for each atom its corresponding ID of the ring(s))
-        void setRingId(size_t atomPos, size_t Id) {            
-			//std::cout << "Pushing Ring ID " << Id << std::endl;
+        void setRingId(int atomPos, size_t Id) {            
+            //std::cout << "Pushing Ring ID " << Id << "atomID" << atomPos << std::endl;
+            
             atoms[atomPos].ringId.push_back(Id);            
         }
